A part of the API has been deprecated in a previous commits, actually it turns out such deprecation is not (fully) needed.
This restores the previous API and cleanup the changes except for hashing strategies and prng which are not user level concerns and would benefit from being in their own packages instead of being top level.
