+++ vertx-auth-common/src/main/java/io/vertx/ext/auth/authorization/AuthorizationContext.java
@@ -14,7 +14,7 @@ package io.vertx.ext.auth.authorization;
 
 import io.vertx.codegen.annotations.VertxGen;
 import io.vertx.core.MultiMap;
-import io.vertx.ext.auth.user.User;
+import io.vertx.ext.auth.User;
 import io.vertx.ext.auth.authorization.impl.AuthorizationContextImpl;
 
 /**
@@ -41,7 +41,7 @@ public interface AuthorizationContext {
    *
    * @return the user
    */
-  io.vertx.ext.auth.user.User user();
+  User user();
 
   /**
    * @return a Multimap containing variable names and values that can be resolved
