+++ vertx-auth-jwt/src/main/java/io/vertx/ext/auth/jwt/JWTAuth.java
@@ -21,7 +21,7 @@ import io.vertx.core.*;
 import io.vertx.core.json.JsonObject;
 import io.vertx.ext.auth.authentication.AuthenticationProvider;
 import io.vertx.ext.auth.jwt.impl.JWTAuthProviderImpl;
-import io.vertx.ext.auth.jose.JWTOptions;
+import io.vertx.ext.auth.JWTOptions;
 
 /**
  * Factory interface for creating JWT based {@link io.vertx.ext.auth.authentication.AuthenticationProvider} instances.
