+++ vertx-auth-oauth2/src/main/java/io/vertx/ext/auth/oauth2/authorization/impl/ScopeAuthorizationImpl.java
@@ -19,7 +19,7 @@ import io.vertx.core.AsyncResult;
 import io.vertx.core.Future;
 import io.vertx.core.Handler;
 import io.vertx.core.json.JsonObject;
-import io.vertx.ext.auth.user.User;
+import io.vertx.ext.auth.User;
 import io.vertx.ext.auth.authorization.Authorization;
 import io.vertx.ext.auth.authorization.PermissionBasedAuthorization;
 import io.vertx.ext.auth.oauth2.authorization.ScopeAuthorization;
